Using unsigned int for useconds_t
The type useconds_t shall be an unsigned integer type capable of storing values at least in the range [0, 1000000]. The type suseconds_t shall be a signed integer type capable of storing values at least in the range [-1, 1000000].

+++ src/lib/evil/evil_unistd.h
@@ -62,7 +62,7 @@ EVIL_API int ftruncate(int fd, off_t size);
 EVIL_API double evil_time_get(void);
 
 #ifdef _MSC_VER
-typedef unsigned long useconds_t;
+typedef unsigned int useconds_t;
 
 EVIL_API void usleep(useconds_t usec);
 #endif
